修复 CI：RUF046 冗余 int() + benchmark 路径错位
Framework CI:
- test_dag_benchmark.py: parents[4] → parents[3]（路径深了 1 级，
  导致找不到 DAG fixture 文件，所有 benchmark 测试 FileNotFoundError）

Game CI:
- curve.py: 移除 int(round(x)) 中多余的 int()（round 已返回 int），
  修复 ruff RUF046 告警

+++ framework/src/calc_framework/inverse/curve.py
@@ -105,8 +105,8 @@ def expand_segment_linear(start: int | float, end: int | float, num_levels: int)
     if num_levels < 1:
         raise ValueError("num_levels 须 >= 1")
     if num_levels == 1:
-        return [int(round(end))]
-    return [int(round(start + (end - start) * i / (num_levels - 1))) for i in range(num_levels)]
+        return [round(end)]
+    return [round(start + (end - start) * i / (num_levels - 1)) for i in range(num_levels)]
 
 
 def _level_overrides_from_params(params: dict[str, Any], spec: SegmentSpec) -> dict[int, float] | None:
